Update
Using a large number of files from different birds highlighted issues:
The normalisation will not fully account for microphone effects. Also
identified and solved a NaN float error , which was due to where a song
is more highly cross-correlated with the padding of the comparison data
than the data itself. This now returns a 0.0% Coefficient.